Refactor Manga model to change rating field to PositiveIntegerField; update serializers and views to accommodate ManyToMany relationship for categories; add script to update manga ratings based on scraping data.

+++ manga/models.py
@@ -33,7 +33,7 @@ class Manga(models.Model):
         verbose_name='カテゴリ'
     )
     description = models.TextField(blank=True, null=True)
-    rating = models.DecimalField(max_digits=3, decimal_places=1, default=0.0)
+    rating = models.PositiveIntegerField(default=0)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
